+++ sermant-plugins/sermant-loadbalancer/config/config.yaml
@@ -1,6 +1,4 @@
 loadbalancer.plugin:
-  #负载均衡配置的key，可自定义
-  key: loadbalancer
   # dubbo负载均衡策略
   # 支持随机（RANDOM，默认），轮询（ROUNDROBIN），最少活跃（LEASTACTIVE），一致性HASH（CONSISTENTHASH），
   # 最短响应时间（SHORTESTRESPONSE，dubbo 2.7.7+）
