--- sermant-plugins/sermant-loadbalancer/dubbo-loadbalancer-plugin/src/main/java/com/huawei/loadbalancer/declarer/UrlDeclarer.java
@@ -1,48 +1,0 @@
-/*
- * Copyright (C) 2022-2022 Huawei Technologies Co., Ltd. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.huawei.loadbalancer.declarer;
-
-import com.huaweicloud.sermant.core.plugin.agent.declarer.AbstractPluginDeclarer;
-import com.huaweicloud.sermant.core.plugin.agent.declarer.InterceptDeclarer;
-import com.huaweicloud.sermant.core.plugin.agent.matcher.ClassMatcher;
-import com.huaweicloud.sermant.core.plugin.agent.matcher.MethodMatcher;
-
-/**
- * URL增强类
- *
- * @author provenceee
- * @since 2022-01-20
- */
-public class UrlDeclarer extends AbstractPluginDeclarer {
-    private static final String[] ENHANCE_CLASS = {"org.apache.dubbo.common.URL", "com.alibaba.dubbo.common.URL"};
-
-    private static final String INTERCEPT_CLASS = "com.huawei.loadbalancer.interceptor.UrlInterceptor";
-
-    private static final String METHOD_NAME = "getMethodParameter";
-
-    @Override
-    public ClassMatcher getClassMatcher() {
-        return ClassMatcher.nameContains(ENHANCE_CLASS);
-    }
-
-    @Override
-    public InterceptDeclarer[] getInterceptDeclarers(ClassLoader classLoader) {
-        return new InterceptDeclarer[]{
-            InterceptDeclarer.build(MethodMatcher.nameEquals(METHOD_NAME), INTERCEPT_CLASS)
-        };
-    }
-}
