+++ sermant-plugins/sermant-loadbalancer/dubbo-loadbalancer-plugin/src/main/resources/META-INF/services/com.huaweicloud.sermant.core.plugin.agent.declarer.PluginDeclarer
@@ -1,1 +1,1 @@
-com.huawei.loadbalancer.declarer.UrlDeclarer
+com.huaweicloud.loadbalancer.declarer.UrlDeclarer
