+++ sermant-plugins/sermant-loadbalancer/loadbalancer-common/src/main/java/com/huaweicloud/loadbalancer/rule/LoadbalancerRuleResolver.java
@@ -33,8 +33,11 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.logging.Logger;
 
 /**
@@ -50,12 +53,27 @@ public class LoadbalancerRuleResolver implements RuleResolver<LoadbalancerRule> 
 
     private static final String MATCH_GROUP_PREFIX = "servicecomb.matchGroup";
 
+    private static final String EMPTY_STR = "";
+
     private final RuleConverter converter;
 
     /**
-     * 规则缓存 key: 业务场景名称 value: 负载均衡规则
+     * 仅存放matchGroup的服务名
+     * <pre>
+     *     key: 业务场景名称
+     *     value: 匹配的服务名
+     * </pre>
      */
-    private final Map<String, LoadbalancerRule> rules = new HashMap<>();
+    private final Map<String, String> serviceCache = new HashMap<>();
+
+    /**
+     * 仅存放loadbalancer的规则类型
+     * <pre>
+     *     key: 业务场景名称
+     *     value: 负载均衡类型
+     * </pre>
+     */
+    private final Map<String, String> ruleCache = new HashMap<>();
 
     /**
      * 服务缓存监听器, 当配置更新时, 同时需要刷新缓存
@@ -63,9 +81,9 @@ public class LoadbalancerRuleResolver implements RuleResolver<LoadbalancerRule> 
     private final List<CacheListener> cacheListeners = new ArrayList<>();
 
     /**
-     * 真实规则数量, 由于存在不匹配的matchGroup, 该数据量仅记录servicecomb.loadbalance的数量
+     * 规则缓存 key: 业务场景名称 value: 负载均衡规则
      */
-    private int realRuleSize;
+    private Map<String, LoadbalancerRule> rules = new ConcurrentHashMap<>();
 
     /**
      * 规则构造器
@@ -77,39 +95,116 @@ public class LoadbalancerRuleResolver implements RuleResolver<LoadbalancerRule> 
     @Override
     public Optional<LoadbalancerRule> resolve(DynamicConfigEvent event) {
         final Optional<LoadbalancerRule> loadbalancerRule = handleConfig(event);
-        loadbalancerRule.ifPresent(rule -> this.notifyCache(rule, event));
+        loadbalancerRule.ifPresent(rule -> {
+            this.notifyCache(rule, event);
+            log(rule, event);
+        });
         return loadbalancerRule;
+    }
+
+    private void log(LoadbalancerRule rule, DynamicConfigEvent event) {
+        if (event.getEventType() == DynamicConfigEventType.DELETE) {
+            LOGGER.info(String.format(Locale.ENGLISH, "[loadbalancer-plugin]Rule [%s] has been deleted!",
+                    rule.toString()));
+        } else if (event.getEventType() == DynamicConfigEventType.MODIFY) {
+            if (rule instanceof ChangedLoadbalancerRule) {
+                final LoadbalancerRule newRule = ((ChangedLoadbalancerRule) rule).getNewRule();
+                final LoadbalancerRule oldRule = ((ChangedLoadbalancerRule) rule).getOldRule();
+                LOGGER.info(String.format(Locale.ENGLISH, "[loadbalancer-plugin]Rule [%s] has changed to [%s]!",
+                        oldRule.toString(), newRule.toString()));
+            }
+        } else {
+            LOGGER.info(String.format(Locale.ENGLISH, "[loadbalancer-plugin]Rule [%s] has been added!",
+                    rule.toString()));
+        }
     }
 
     private Optional<LoadbalancerRule> handleConfig(DynamicConfigEvent event) {
         final String key = event.getKey();
-        if (key.startsWith(LOAD_BALANCER_PREFIX)) {
-            return handleRule(event);
-        } else if (key.startsWith(MATCH_GROUP_PREFIX)) {
-            return handleMatchGroup(event);
-        } else {
+        if (!isTargetConfig(key)) {
             return Optional.empty();
         }
+        if (key.startsWith(LOAD_BALANCER_PREFIX)) {
+            handleRule(event);
+        } else {
+            handleMatchGroup(event);
+        }
+        return combine();
     }
 
-    private Optional<LoadbalancerRule> handleMatchGroup(DynamicConfigEvent event) {
+    private boolean isTargetConfig(String key) {
+        return key != null && (key.startsWith(LOAD_BALANCER_PREFIX) || key.startsWith(MATCH_GROUP_PREFIX));
+    }
+
+    /**
+     * 组合matchGroup serviceName与loadbalancer rule,并返回变更的规则
+     */
+    private Optional<LoadbalancerRule> combine() {
+        final Map<String, LoadbalancerRule> newRules = new ConcurrentHashMap<>(ruleCache.size());
+        final Set<Entry<String, String>> entries = ruleCache.entrySet();
+        for (Entry<String, String> entry : entries) {
+            final String serviceName = serviceCache.get(entry.getKey());
+            if (serviceName == null || entry.getValue() == null) {
+                continue;
+            }
+            newRules.put(entry.getKey(), new LoadbalancerRule(EMPTY_STR.equals(serviceName) ? null : serviceName,
+                    entry.getValue()));
+        }
+        final Map<String, LoadbalancerRule> oldRules = new HashMap<>(this.rules);
+        this.rules = newRules;
+        return getChangedRule(oldRules, newRules);
+    }
+
+    /**
+     * 获取变更的规则:
+     *
+     * <p>删除: 返回被删除的规则</p>
+     * <p>新增: 返回新增的规则</p>
+     * <p>更改: 返回更改前后的规则, 见{@link ChangedLoadbalancerRule}</p>
+     *
+     * @param oldRules 旧规则集合
+     * @param newRules 新规则集合
+     * @return 变更的规则
+     */
+    private Optional<LoadbalancerRule> getChangedRule(Map<String, LoadbalancerRule> oldRules,
+            Map<String, LoadbalancerRule> newRules) {
+        for (Entry<String, LoadbalancerRule> entry : oldRules.entrySet()) {
+            final LoadbalancerRule rule = newRules.get(entry.getKey());
+            if (rule == null) {
+                // 已删除的规则
+                return Optional.ofNullable(entry.getValue());
+            }
+            if (!StringUtils.equals(entry.getValue().getRule(), rule.getRule())
+                    || !StringUtils.equals(entry.getValue().getServiceName(), rule.getServiceName())) {
+                // 变更后的规则
+                return Optional.of(new ChangedLoadbalancerRule(entry.getValue(), rule));
+            }
+        }
+
+        // 新增规则
+        if (newRules.size() > oldRules.size()) {
+            for (Entry<String, LoadbalancerRule> entry : newRules.entrySet()) {
+                if (oldRules.get(entry.getKey()) == null) {
+                    return Optional.ofNullable(entry.getValue());
+                }
+            }
+        }
+        return Optional.empty();
+    }
+
+    private void handleMatchGroup(DynamicConfigEvent event) {
         final String businessKey = event.getKey().substring(MATCH_GROUP_PREFIX.length() + 1);
         if (event.getEventType() == DynamicConfigEventType.DELETE) {
-            return Optional.ofNullable(rules.remove(businessKey));
+            serviceCache.remove(businessKey);
+            return;
         }
         final Optional<Map> convert = converter.convert(event.getContent(), Map.class);
         if (!convert.isPresent()) {
-            return Optional.empty();
+            return;
         }
-        LoadbalancerRule rule = rules.getOrDefault(businessKey, new LoadbalancerRule());
         final Map<String, Object> matcher = convert.get();
         final Optional<String> serviceNameOptional = resolveServiceName(matcher);
-        if (!serviceNameOptional.isPresent()) {
-            return Optional.of(rule);
-        }
-        rule.setServiceName(serviceNameOptional.get());
-        rules.put(businessKey, rule);
-        return Optional.of(rule);
+        serviceCache.put(businessKey, serviceNameOptional.orElse(EMPTY_STR));
     }
 
     private Optional<String> resolveServiceName(Map<String, Object> matcher) {
@@ -117,35 +212,28 @@ public class LoadbalancerRuleResolver implements RuleResolver<LoadbalancerRule> 
         if (matches instanceof List) {
             final List<Map<String, Object>> list = (List<Map<String, Object>>) matches;
             if (!list.isEmpty()) {
-                return Optional.ofNullable((String) list.get(0).get("serviceName"));
+                final Object serviceName = list.get(0).get("serviceName");
+                return Optional.ofNullable(serviceName == null ? EMPTY_STR : String.valueOf(serviceName));
             }
         }
         return Optional.empty();
     }
 
-    private Optional<LoadbalancerRule> handleRule(DynamicConfigEvent event) {
+    private void handleRule(DynamicConfigEvent event) {
         String businessKey = event.getKey().substring(LOAD_BALANCER_PREFIX.length() + 1);
         if (event.getEventType() == DynamicConfigEventType.DELETE) {
-            final LoadbalancerRule remove = rules.remove(businessKey);
-            if (remove != null) {
-                realRuleSize--;
-            }
-            return Optional.ofNullable(remove);
+            ruleCache.remove(businessKey);
+            return;
         }
-        final LoadbalancerRule rule = rules.getOrDefault(businessKey, new LoadbalancerRule());
         final Optional<LoadbalancerRule> ruleOptional = converter.convert(event.getContent(), LoadbalancerRule.class);
-        if (ruleOptional.isPresent()) {
-            final String loadbalancerType = ruleOptional.get().getRule();
-            if (isSupport(loadbalancerType)) {
-                rule.setRule(loadbalancerType);
-                rules.put(businessKey, rule);
-                realRuleSize++;
+        ruleOptional.ifPresent(rule -> {
+            if (isSupport(rule.getRule())) {
+                ruleCache.put(businessKey, rule.getRule());
             } else {
-                LOGGER.warning(String.format(Locale.ENGLISH, "Can not support loadbalancer rule [%s]",
-                        loadbalancerType));
+                LOGGER.warning(String.format(Locale.ENGLISH, "Can not support loadbalancer rule: [%s]",
+                        rule.getRule()));
             }
-        }
-        return Optional.of(rule);
+        });
     }
 
     private boolean isSupport(String loadbalancerType) {
@@ -171,12 +259,12 @@ public class LoadbalancerRuleResolver implements RuleResolver<LoadbalancerRule> 
     }
 
     /**
-     * 宿主服务是否配置负载均衡策略
+     * 判断当前的缓存是否存在, 若存在配置内容则为已配置
      *
-     * @return true为已配置, 根据实际规则数量判断, 数量大于0则为已配置
+     * @return true为已配置
      */
     public boolean isConfigured() {
-        return realRuleSize > 0;
+        return !ruleCache.isEmpty() || !serviceCache.isEmpty();
     }
 
     /**
