+++ sermant-plugins/sermant-loadbalancer/spring-cloud-loadbalancer-plugin/src/main/resources/META-INF/services/com.huaweicloud.sermant.core.plugin.agent.declarer.PluginDeclarer
@@ -1,5 +1,4 @@
-com.huawei.loadbalancer.declarer.ClientFactoryDeclarer
-com.huawei.loadbalancer.declarer.RandomDeclarer
-com.huawei.loadbalancer.declarer.RibbonLoadBalancerDeclarer
-com.huawei.loadbalancer.declarer.RoundRobinDeclarer
-com.huawei.loadbalancer.declarer.RibbonClientDeclarer
+com.huaweicloud.loadbalancer.declarer.ClientFactoryDeclarer
+com.huaweicloud.loadbalancer.declarer.RandomDeclarer
+com.huaweicloud.loadbalancer.declarer.RibbonLoadBalancerDeclarer
+com.huaweicloud.loadbalancer.declarer.RoundRobinDeclarer
